# 原型/lumina-api-openapi.yaml
openapi: 3.1.0
info:
  title: Lumina 墨光 · API
  version: 1.5.0
  description: |
    Lumina（墨光）教育应用 RESTful API，覆盖用户、课程、作业、成绩、直播、协作、AI 等模块。

    ## 特性
    - JWT 双令牌认证（Access Token 15min + Refresh Token 7d）
    - Cursor-based 分页
    - RFC 7807 错误格式
    - SSE 流式响应（AI 接口）
    - 直播课堂：REST + after_id 轮询（聊天/举手/答题/点名）· HLS 流媒体（LIVE_STREAM_* 可插拔，未接入返回 mock://）
    - 小组协作：课程小组 / 成员 / 项目 / 看板拖拽 / 共享文件 / 组内讨论（D-02）
    - 账户补全：自助注册（POST /auth/register）+ 消息通知中心（/notifications/my · D-03）

  contact:
    name: Lumina Team
    url: https://github.com/yuan50697105/lumina-edu
  license:
    name: MIT
servers:
  - url: https://api.lumina.edu/api/v1
    description: 生产环境
  - url: https://api-staging.lumina.edu/api/v1
    description: 预发布环境
  - url: http://localhost:8080/api/v1
    description: 本地开发
...
security:
  - BearerAuth: []

# ─── 路径 ───
paths:
...
  /courses:
    get:
      tags: [courses]
      summary: 课程列表
      parameters:
        - $ref: '#/components/parameters/Semester'
        - $ref: '#/components/parameters/Department'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 0
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Course'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
    post:
      tags: [courses]
      summary: 创建课程
      description: 教师创建新课程
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseCreateRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 0
                  data:
                    $ref: '#/components/schemas/Course'
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT Access Token（15 分钟有效）"

  # ─── 通用参数 ───
  parameters:
...
    Semester:
      name: semester
      in: query
      schema:
        type: string
        example: "2026-1"
      description: "学期筛选"
    Department:
      name: department
      in: query
      schema:
        type: string
      description: "院系筛选"
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
      description: "分页游标"
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: "每页数量（默认 20，最大 100）"

  # ─── 通用响应 ───
...
  schemas:
...
    Pagination:
      type: object
      properties:
        cursor:
          type: string
        has_more:
          type: boolean
        total:
          type: integer
...
    Course:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
          example: "CS201"
        title:
          type: string
          example: "高等数学"
        description:
          type: string
        teacher:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
        credits:
          type: number
          example: 4.0
        students_count:
          type: integer
          example: 86
        semester:
          type: string
          example: "2026-1"
        department:
          type: string
        schedule:
          type: array
          items:
            type: object
            properties:
              day:
                type: integer
                description: "1=周一"
              start:
                type: string
                example: "08:00"
              end:
                type: string
                example: "09:40"
              room:
                type: string
        created_at:
          type: string
          format: date-time
    CourseCreateRequest:
      type: object
      required: [title, code, credits]
      properties:
        title:
          type: string
        code:
          type: string
        description:
          type: string
        credits:
          type: number
        department:
          type: string
        schedule:
          type: array
          items:
            type: object
            properties:
              day:
                type: integer
              start:
                type: string
              end:
                type: string
              room:
                type: string
...